+++ Sources/OmFileFormatC/src/om_common.c
@@ -13,8 +13,8 @@
 #include "fp.h"
 
 /// Assume chunk buffer is a 16 bit integer array and convert to float
-void om_common_copy_float_to_int16(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy_float_to_int16(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         float val = ((float *)src)[i];
         if (isnan(val)) {
             ((int16_t *)dst)[i] = INT16_MAX;
@@ -26,8 +26,8 @@ void om_common_copy_float_to_int16(uint64_t length, float scale_factor, float ad
 }
 
 /// Assume chunk buffer is a 16 bit integer array and convert to float and scale log10
-void om_common_copy_float_to_int16_log10(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy_float_to_int16_log10(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         float val = ((float *)src)[i];
         if (isnan(val)) {
             ((float *)dst)[i] = INT16_MAX;
@@ -39,58 +39,58 @@ void om_common_copy_float_to_int16_log10(uint64_t length, float scale_factor, fl
 }
 
 /// Assume chunk buffer is a 16 bit integer array and convert to float
-void om_common_copy_int16_to_float(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy_int16_to_float(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         int16_t val = ((int16_t *)src)[i];
         ((float *)dst)[i] = (val == INT16_MAX) ? NAN : (float)val / scale_factor - add_offset;
     }
 }
 
 /// Assume chunk buffer is a 16 bit integer array and convert to float and scale log10
-void om_common_copy_int16_to_float_log10(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy_int16_to_float_log10(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         int16_t val = ((int16_t *)src)[i];
         ((float *)dst)[i] = (val == INT16_MAX) ? NAN : powf(10, (float)val / scale_factor) - 1;
     }
 }
 
-void om_common_copy8(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy8(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         ((int8_t *)dst)[i] = ((int8_t *)src)[i];
     }
 }
 
-void om_common_copy16(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy16(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         ((int16_t *)dst)[i] = ((int16_t *)src)[i];
     }
 }
 
-void om_common_copy32(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy32(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         ((int32_t *)dst)[i] = ((int32_t *)src)[i];
     }
 }
 
-void om_common_copy64(uint64_t length, float scale_factor, float add_offset, const void* src, void* dst) {
-    for (uint64_t i = 0; i < length; ++i) {
+void om_common_copy64(size_t length, float scale_factor, float add_offset, const void* src, void* dst) {
+    for (size_t i = 0; i < length; ++i) {
         ((int64_t *)dst)[i] = ((int64_t *)src)[i];
     }
 }
 
-uint64_t om_common_compress_fpxenc32(const void* src, uint64_t length, void* dst) {
+size_t om_common_compress_fpxenc32(const void* src, size_t length, void* dst) {
     return fpxenc32((uint32_t*)src, length, (unsigned char *)dst, 0);
 }
 
-uint64_t om_common_compress_fpxenc64(const void* src, uint64_t length, void* dst) {
+size_t om_common_compress_fpxenc64(const void* src, size_t length, void* dst) {
     return fpxenc64((uint64_t*)src, length, (unsigned char *)dst, 0);
 }
 
-uint64_t om_common_decompress_fpxdec32(const void* src, uint64_t length, void* dst) {
+size_t om_common_decompress_fpxdec32(const void* src, size_t length, void* dst) {
     return fpxdec32((unsigned char *)src, length, (uint32_t *)dst, 0);
 }
 
-uint64_t om_common_decompress_fpxdec64(const void* src, uint64_t length, void* dst) {
+size_t om_common_decompress_fpxdec64(const void* src, size_t length, void* dst) {
     return fpxdec64((unsigned char *)src, length, (uint64_t *)dst, 0);
 }
